Implement label reuse via reverse queue
Now, when label is complitely finished by ComputationBlock, it gets
enqueued on the reverseQueue_ of LabellingBlock. Then LabellingBlock
dequeued it and puts into his LabelSet for later reuse.

Now time determinism between Blocks is lost. Label is not guaranteed
to be reused right after it was pushed on the reverseQueue. Because
of that tests, which dependent on the label number, are broken.